Adds selector stub activity.
