+++ app/src/main/res/layout/activity_main.xml
@@ -1,6 +1,7 @@
-<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+<android.support.v4.widget.DrawerLayout
     android:id="@+id/drawer_layout"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
@@ -10,34 +11,25 @@
         android:layout_height="match_parent"
         android:scaleType="centerCrop"
         android:src="@drawable/scalac_bg"
-        tools:ignore="ContentDescription" />
+        tools:ignore="ContentDescription"/>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical">
 
-        <include layout="@layout/toolbar_with_spinner" />
+        <include layout="@layout/toolbar_with_spinner"/>
 
         <FrameLayout
             android:id="@+id/content_frame"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <ProgressBar
-                android:id="@+id/homeProgressBar"
-                style="?android:attr/progressBarStyleLarge"
-                android:layout_width="32dp"
-                android:layout_height="32dp"
-                android:layout_gravity="center" />
-
-        </FrameLayout>
+            android:layout_height="match_parent"/>
     </LinearLayout>
 
     <include
         layout="@layout/drawer_navigation"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:layout_gravity="start" />
+        android:layout_gravity="start"/>
 
 </android.support.v4.widget.DrawerLayout>
